+++ doc/about/changes.md
@@ -4,10 +4,14 @@
 in progress
 {: .subtitle}
 
+### Breaking changes
+- Require Python 3.6+
+
 ### Bug fixes
 - Error reporting now shows the proper stack frames on Windows
 - Fix detection of `clang-cl` compiler
 - MinGW can now use libraries of the form `name.lib`
+- Replace `packaging` dependency with `verspec` for future stability
 
 ---